[12/04/17] Moving to making it run.
AnimalDevelopment, AnimalStrain, and AnimalStrain are modified to
protect them from modification during the simulation. Needs to have
improved methods for altering the properties of these through the zoo.

Includes a new Temporary class which will be used to provide default
options whilst the input.txt file format is developed. It brings in a
default zoo for running simulations directly from the code.

Adds SimulationOptions2 as the basis for a new more flexible
SimulationOptions object. But it requires additional work. This is
supported by an additional SimulationOptionSetting object for providing
those settings.

Includes some unused changes to StringFormula which should be in future
extracted using the ReadableMath package under development elsewhere.