[19/04/17] SimulationOptions2 Revisited
Significant improvements are made to the SimulationOptions2 object but
are not yet fully operable. Modifications to AnimalZoo are made with the
beginnings of reading the zoo from the input.txt file, but have not yet
been completed. Context is temporarily improved but does not yet ready.
Javadoc is starting to be revised.