Delayed Update 6/6/17
Replaces a number of tracked values to a new tracked system allowing for
details of mean and variance to be extracted. This can be expanded to a
large number of other trackings now. There needs to be absolute
verification that the values are being calculated correctly. Launches
from code correctly and appears to optimise correctly, but the
optimisation rate is very slow with present settings. Further tweaking
is required for improving rate of optimisation. Future: Produce Linear
reaction optimisations and more algorithmic ones.